把ethernetif单板相关代码剥离
