+++ components/net/lwip_port/OS/ethernetif.c
@@ -77,20 +77,19 @@
 * something that better describes your network interface.
 */
 
-#include "stm32f4xx_hal.h"
-#include "stm32f4xx_hal_eth.h"
-#include "lwip/opt.h"
+
 #include "lwip/def.h"
 #include "lwip/mem.h"
-#include "lwip/pbuf.h"
 #include "lwip/timeouts.h"
 #include "netif/etharp.h"
-#include "lwip/err.h"
 #include "ethernetif.h"
 #include <string.h>
 
 #define netifINTERFACE_TASK_STACK_SIZE              ( 4096u )
 #define netifINTERFACE_TASK_PRIORITY                ( 6u   )
+/* The time to block waiting for input. */
+#define TIME_WAITING_FOR_INPUT                      ( 500u )
+
 
 /* Define those to better describe your network interface. */
 #define IFNAME0 's'
@@ -98,169 +97,31 @@
 
 
 static struct netif* s_pxNetIf = NULL;
-
-/* Private variables ---------------------------------------------------------*/
-#if defined ( __ICCARM__ ) /*!< IAR Compiler */
-#pragma data_alignment=4
-#endif
-__ALIGN_BEGIN ETH_DMADescTypeDef  DMARxDscrTab[ETH_RXBUFNB] __ALIGN_END;/* Ethernet Rx MA Descriptor */
-
-#if defined ( __ICCARM__ ) /*!< IAR Compiler */
-#pragma data_alignment=4
-#endif
-__ALIGN_BEGIN ETH_DMADescTypeDef  DMATxDscrTab[ETH_TXBUFNB] __ALIGN_END;/* Ethernet Tx DMA Descriptor */
-
-#if defined ( __ICCARM__ ) /*!< IAR Compiler */
-#pragma data_alignment=4
-#endif
-__ALIGN_BEGIN uint8_t Rx_Buff[ETH_RXBUFNB][ETH_RX_BUF_SIZE] __ALIGN_END; /* Ethernet Receive Buffer */
-
-#if defined ( __ICCARM__ ) /*!< IAR Compiler */
-#pragma data_alignment=4
-#endif
-__ALIGN_BEGIN uint8_t Tx_Buff[ETH_TXBUFNB][ETH_TX_BUF_SIZE] __ALIGN_END; /* Ethernet Transmit Buffer */
-
-/* Global Ethernet handle */
-ETH_HandleTypeDef heth;
-
-/* The time to block waiting for input. */
-#define TIME_WAITING_FOR_INPUT                 ( 500 )
+static struct eth_api s_eth_api; 
 
 /* Semaphore to signal incoming packets */
-sys_sem_t s_xSemaphore;
+static sys_sem_t s_xSemaphore;
 
-/**
-  * @brief  Ethernet Rx Transfer completed callback
-  * @param  heth: ETH handle
-  * @retval None
-  */
-void HAL_ETH_RxCpltCallback(ETH_HandleTypeDef *pheth)
-{
-	sys_sem_signal(&s_xSemaphore);
-}
 
-/* Os ethernet interrupt handler */
-void ETH_IRQHandler(void)
-{
-    HAL_ETH_IRQHandler(&heth);
-}
 //void ethernetif_input( void * pvParameters );
 static void arp_timer(void* arg);
 
-
-/* Private functions ---------------------------------------------------------*/
-
-void HAL_ETH_MspInit(ETH_HandleTypeDef* ethHandle)
+int8_t ethernetif_api_register(struct eth_api *api)
 {
-    GPIO_InitTypeDef GPIO_InitStruct;
-
-    __HAL_RCC_GPIOC_CLK_ENABLE();
-    __HAL_RCC_GPIOH_CLK_ENABLE();
-    __HAL_RCC_GPIOA_CLK_ENABLE();
-    __HAL_RCC_GPIOB_CLK_ENABLE();
-    __HAL_RCC_GPIOG_CLK_ENABLE();
-    __HAL_RCC_GPIOI_CLK_ENABLE();
-
-
-    if (ethHandle->Instance == ETH)
+    if(api == NULL) 
     {
-        /* USER CODE BEGIN ETH_MspInit 0 */
-
-        /* USER CODE END ETH_MspInit 0 */
-        /* Enable Peripheral clock */
-        __HAL_RCC_ETH_CLK_ENABLE();
-
-        /**ETH GPIO Configuration
-        PC1     ------> ETH_MDC
-        PA1     ------> ETH_REF_CLK
-        PA2     ------> ETH_MDIO
-        PA7     ------> ETH_CRS_DV
-        PC4     ------> ETH_RXD0
-        PC5     ------> ETH_RXD1
-        PB11     ------> ETH_TX_EN
-        PG13     ------> ETH_TXD0
-        PG14     ------> ETH_TXD1
-            PI1      ------> ETH_NRST
-        */
-
-        GPIO_InitStruct.Pin = GPIO_PIN_1;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-        HAL_GPIO_Init(GPIOI, &GPIO_InitStruct);
-
-        HAL_GPIO_WritePin(GPIOI, GPIO_PIN_1, GPIO_PIN_RESET);
-        HAL_Delay(1000);
-        HAL_GPIO_WritePin(GPIOI, GPIO_PIN_1, GPIO_PIN_SET);
-        HAL_Delay(1000);
-
-        GPIO_InitStruct.Pin = GPIO_PIN_1 | GPIO_PIN_4 | GPIO_PIN_5;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-        GPIO_InitStruct.Alternate = GPIO_AF11_ETH;
-        HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-        GPIO_InitStruct.Pin = GPIO_PIN_1 | GPIO_PIN_2 | GPIO_PIN_7;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-        GPIO_InitStruct.Alternate = GPIO_AF11_ETH;
-        HAL_GPIO_Init(GPIOA, &GPIO_InitStruct);
-
-        GPIO_InitStruct.Pin = GPIO_PIN_11;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-        GPIO_InitStruct.Alternate = GPIO_AF11_ETH;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        GPIO_InitStruct.Pin = GPIO_PIN_13 | GPIO_PIN_14;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_VERY_HIGH;
-        GPIO_InitStruct.Alternate = GPIO_AF11_ETH;
-        HAL_GPIO_Init(GPIOG, &GPIO_InitStruct);
-        /* USER CODE BEGIN ETH_MspInit 1 */
-
-        /* USER CODE END ETH_MspInit 1 */
+        return ERR_ARG;
     }
+    memcpy(&s_eth_api, api, sizeof(struct eth_api));
+
+    return ERR_OK;
 }
 
-void HAL_ETH_MspDeInit(ETH_HandleTypeDef* ethHandle)
+
+void ethernetif_rxcb(void)
 {
-    if (ethHandle->Instance == ETH)
-    {
-        /* USER CODE BEGIN ETH_MspDeInit 0 */
-
-        /* USER CODE END ETH_MspDeInit 0 */
-        /* Peripheral clock disable */
-        __HAL_RCC_ETH_CLK_DISABLE();
-
-        /**ETH GPIO Configuration
-        PC1     ------> ETH_MDC
-        PA1     ------> ETH_REF_CLK
-        PA2     ------> ETH_MDIO
-        PA7     ------> ETH_CRS_DV
-        PC4     ------> ETH_RXD0
-        PC5     ------> ETH_RXD1
-        PB11     ------> ETH_TX_EN
-        PG13     ------> ETH_TXD0
-        PG14     ------> ETH_TXD1
-        */
-        HAL_GPIO_DeInit(GPIOC, GPIO_PIN_1 | GPIO_PIN_4 | GPIO_PIN_5);
-
-        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_1 | GPIO_PIN_2 | GPIO_PIN_7);
-
-        HAL_GPIO_DeInit(GPIOB, GPIO_PIN_11);
-        HAL_GPIO_DeInit(GPIOG, GPIO_PIN_13 | GPIO_PIN_14);
-
-        /* USER CODE BEGIN ETH_MspDeInit 1 */
-
-        /* USER CODE END ETH_MspDeInit 1 */
-    }
+    sys_sem_signal(&s_xSemaphore);
 }
-
 
 /**
 * In this function, the hardware should be initialized.
@@ -271,58 +132,18 @@ void HAL_ETH_MspDeInit(ETH_HandleTypeDef* ethHandle)
 */
 static void low_level_init(struct netif* netif)
 {
-    HAL_StatusTypeDef hal_eth_init_status;
-
-    /* Init ETH */
-
-    uint8_t MACAddr[6] ;
-    heth.Instance = ETH;
-    heth.Init.AutoNegotiation = ETH_AUTONEGOTIATION_ENABLE;
-    heth.Init.PhyAddress = LAN8742A_PHY_ADDRESS;
-    MACAddr[0] = 0x00;
-    MACAddr[1] = 0x80;
-    MACAddr[2] = 0xE1;
-    MACAddr[3] = 0x00;
-    MACAddr[4] = 0x00;
-    MACAddr[5] = 0x00;
-    heth.Init.MACAddr = &MACAddr[0];
-    heth.Init.RxMode = ETH_RXINTERRUPT_MODE;
-    heth.Init.ChecksumMode = ETH_CHECKSUM_BY_HARDWARE;
-    heth.Init.MediaInterface = ETH_MEDIA_INTERFACE_RMII;
-
-    /* USER CODE BEGIN MACADDRESS */
-
-    /* USER CODE END MACADDRESS */
-
-    hal_eth_init_status = HAL_ETH_Init(&heth);
-
-    if (hal_eth_init_status == HAL_OK)
-    {
-        /* Set netif link flag */
-        netif->flags |= NETIF_FLAG_LINK_UP;
-    }
-
-    /* Initialize Tx Descriptors list: Chain Mode */
-    (void)HAL_ETH_DMATxDescListInit(&heth, DMATxDscrTab, &Tx_Buff[0][0], ETH_TXBUFNB);
-
-    /* Initialize Rx Descriptors list: Chain Mode  */
-    (void)HAL_ETH_DMARxDescListInit(&heth, DMARxDscrTab, &Rx_Buff[0][0], ETH_RXBUFNB);
+    
+    if(s_eth_api.init!=NULL)
+        s_eth_api.init(netif);
 
 #if LWIP_ARP || LWIP_ETHERNET
 
     /* set MAC hardware address length */
     netif->hwaddr_len = ETH_HWADDR_LEN;
-
-    /* set MAC hardware address */
-    netif->hwaddr[0] =  heth.Init.MACAddr[0];
-    netif->hwaddr[1] =  heth.Init.MACAddr[1];
-    netif->hwaddr[2] =  heth.Init.MACAddr[2];
-    netif->hwaddr[3] =  heth.Init.MACAddr[3];
-    netif->hwaddr[4] =  heth.Init.MACAddr[4];
-    netif->hwaddr[5] =  heth.Init.MACAddr[5];
-
     /* maximum transfer unit */
     netif->mtu = 1500;
+    if(s_eth_api.set_mac)
+        s_eth_api.set_mac(netif);
 
     /* Accept broadcast address and ARP traffic */
     /* don't set NETIF_FLAG_ETHARP if this device is not an ethernet one */
@@ -336,13 +157,13 @@ static void low_level_init(struct netif* netif)
 
     if (ERR_OK != sys_sem_new(&s_xSemaphore, 1))
     {
-	return;    
+	    return;    
     }
     /* create the task that handles the ETH_MAC */
     sys_thread_new((char*)"Eth_if", ethernetif_input, netif, netifINTERFACE_TASK_STACK_SIZE, netifINTERFACE_TASK_PRIORITY);
 
-    /* Enable MAC and DMA transmission and reception */
-    (void)HAL_ETH_Start(&heth);
+    if(s_eth_api.start)
+        s_eth_api.start(netif);
 
 #endif /* LWIP_ARP || LWIP_ETHERNET */
 
@@ -365,79 +186,14 @@ static void low_level_init(struct netif* netif)
 */
 static err_t low_level_output(struct netif* netif, struct pbuf* p)
 {
-    err_t errval;
-    struct pbuf* q;
-    uint8_t* buffer = (uint8_t*)(heth.TxDesc->Buffer1Addr);
-    __IO ETH_DMADescTypeDef* DmaTxDesc;
-    uint32_t framelength = 0;
-    uint32_t bufferoffset = 0;
-    uint32_t byteslefttocopy = 0;
-    uint32_t payloadoffset = 0;
-    DmaTxDesc = heth.TxDesc;
-    bufferoffset = 0;
-
-    /* copy frame from pbufs to driver buffers */
-    for (q = p; q != NULL; q = q->next)
+    err_t err = ERR_IF;
+    
+    if(s_eth_api.output)
     {
-        /* Is this buffer available? If not, goto error */
-        if ((DmaTxDesc->Status & ETH_DMATXDESC_OWN) != (uint32_t)RESET)
-        {
-            errval = ERR_USE;
-            goto error;
-        }
-
-        /* Get bytes in current lwIP buffer */
-        byteslefttocopy = q->len;
-        payloadoffset = 0;
-
-        /* Check if the length of data to copy is bigger than Tx buffer size*/
-        while ( (byteslefttocopy + bufferoffset) > ETH_TX_BUF_SIZE )
-        {
-            /* Copy data to Tx buffer*/
-            memcpy( (uint8_t*)((uint8_t*)buffer + bufferoffset), (uint8_t*)((uint8_t*)q->payload + payloadoffset), (ETH_TX_BUF_SIZE - bufferoffset) );
-
-            /* Point to next descriptor */
-            DmaTxDesc = (ETH_DMADescTypeDef*)(DmaTxDesc->Buffer2NextDescAddr);
-
-            /* Check if the buffer is available */
-            if ((DmaTxDesc->Status & ETH_DMATXDESC_OWN) != (uint32_t)RESET)
-            {
-                errval = ERR_USE;
-                goto error;
-            }
-
-            buffer = (uint8_t*)(DmaTxDesc->Buffer1Addr);
-
-            byteslefttocopy = byteslefttocopy - (ETH_TX_BUF_SIZE - bufferoffset);
-            payloadoffset = payloadoffset + (ETH_TX_BUF_SIZE - bufferoffset);
-            framelength = framelength + (ETH_TX_BUF_SIZE - bufferoffset);
-            bufferoffset = 0;
-        }
-
-        /* Copy the remaining bytes */
-        memcpy( (uint8_t*)((uint8_t*)buffer + bufferoffset), (uint8_t*)((uint8_t*)q->payload + payloadoffset), byteslefttocopy );
-        bufferoffset = bufferoffset + byteslefttocopy;
-        framelength = framelength + byteslefttocopy;
+        err = s_eth_api.output(netif, p);
     }
 
-    /* Prepare transmit descriptors to give to DMA */
-    (void)HAL_ETH_TransmitFrame(&heth, framelength);
-
-    errval = ERR_OK;
-
-error:
-
-    /* When Transmit Underflow flag is set, clear it and issue a Transmit Poll Demand to resume transmission */
-    if ((heth.Instance->DMASR & ETH_DMASR_TUS) != (uint32_t)RESET)
-    {
-        /* Clear TUS ETHERNET DMA flag */
-        heth.Instance->DMASR = ETH_DMASR_TUS;
-
-        /* Resume DMA transmission*/
-        heth.Instance->DMATPDR = 0;
-    }
-
-    return errval;
+    return err;
 }
 
 /**
@@ -450,88 +206,13 @@ error:
 */
 static struct pbuf* low_level_input(struct netif* netif)
 {
-    struct pbuf* p = NULL;
-    struct pbuf* q = NULL;
-    uint16_t len = 0;
-    uint8_t* buffer;
-    __IO ETH_DMADescTypeDef* dmarxdesc;
-    uint32_t bufferoffset = 0;
-    uint32_t payloadoffset = 0;
-    uint32_t byteslefttocopy = 0;
-    uint32_t i = 0;
-
-
-    /* get received frame */
-    if (HAL_ETH_GetReceivedFrame(&heth) != HAL_OK)
-    { 
-	    return NULL;
-    }
-
-    /* Obtain the size of the packet and put it into the "len" variable. */
-    len = heth.RxFrameInfos.length;
-    buffer = (uint8_t*)heth.RxFrameInfos.buffer;
-
-    if (len > 0)
+    struct pbuf *p = NULL;
+    if(s_eth_api.input)
     {
-        /* We allocate a pbuf chain of pbufs from the Lwip buffer pool */
-        p = pbuf_alloc(PBUF_RAW, len, PBUF_POOL);
+        p = s_eth_api.input(netif);
     }
-
-    if (p != NULL)
-    {
-        dmarxdesc = heth.RxFrameInfos.FSRxDesc;
-        bufferoffset = 0;
-
-        for (q = p; q != NULL; q = q->next)
-        {
-            byteslefttocopy = q->len;
-            payloadoffset = 0;
-
-            /* Check if the length of bytes to copy in current pbuf is bigger than Rx buffer size*/
-            while ( (byteslefttocopy + bufferoffset) > ETH_RX_BUF_SIZE )
-            {
-                /* Copy data to pbuf */
-                memcpy( (uint8_t*)((uint8_t*)q->payload + payloadoffset), (uint8_t*)((uint8_t*)buffer + bufferoffset), (ETH_RX_BUF_SIZE - bufferoffset));
-
-                /* Point to next descriptor */
-                dmarxdesc = (ETH_DMADescTypeDef*)(dmarxdesc->Buffer2NextDescAddr);
-                buffer = (uint8_t*)(dmarxdesc->Buffer1Addr);
-
-                byteslefttocopy = byteslefttocopy - (ETH_RX_BUF_SIZE - bufferoffset);
-                payloadoffset = payloadoffset + (ETH_RX_BUF_SIZE - bufferoffset);
-                bufferoffset = 0;
-            }
-
-            /* Copy remaining data in pbuf */
-            memcpy( (uint8_t*)((uint8_t*)q->payload + payloadoffset), (uint8_t*)((uint8_t*)buffer + bufferoffset), byteslefttocopy);
-            bufferoffset = bufferoffset + byteslefttocopy;
-        }
-    }
-
-    /* Release descriptors to DMA */
-    /* Point to first descriptor */
-    dmarxdesc = heth.RxFrameInfos.FSRxDesc;
-
-    /* Set Own bit in Rx descriptors: gives the buffers back to DMA */
-    for (i = 0; i < heth.RxFrameInfos.SegCount; i++)
-    {
-        dmarxdesc->Status |= ETH_DMARXDESC_OWN;
-        dmarxdesc = (ETH_DMADescTypeDef*)(dmarxdesc->Buffer2NextDescAddr);
-    }
-
-    /* Clear Segment_Count */
-    heth.RxFrameInfos.SegCount = 0;
-
-    /* When Rx Buffer unavailable flag is set: clear it and resume reception */
-    if ((heth.Instance->DMASR & ETH_DMASR_RBUS) != (uint32_t)RESET)
-    {
-        /* Clear RBUS ETHERNET DMA flag */
-        heth.Instance->DMASR = ETH_DMASR_RBUS;
-        /* Resume DMA reception */
-        heth.Instance->DMARPDR = 0;
-    }
-
     return p;
+    
 }
 
 /**
