+++ targets/STM32F429IGTx_FIRE/Inc/eth.h
@@ -1,0 +1,12 @@
+#ifndef __STM32F429_ETH_H__
+#define __STM32F429_ETH_H__
+
+#include "ethernetif.h"
+
+#include "stm32f4xx_hal.h"
+#include "stm32f4xx_hal_eth.h"
+
+
+extern struct eth_api g_eth_api;
+
+#endif
