+++ targets/STM32F429IGTx_FIRE/Inc/sys_init.h
@@ -72,6 +72,7 @@
 
 #include "net.h"
 #include "ssl.h"
+#include "eth.h"
 
 #ifdef __cplusplus
  extern "C" {
